More accurate flushing for sea side
When the doors are open for a long period of time on the sea side, and
one is flushing fresh water through the lock, the lock chamber will not
necessarily end up with the salinity of the lake. Only if one flushes
with a very high velocity (> velocity of density current) does this
happen. When flushing with a lower velocity, the density current can
still enter the lock chamber, and one ends up with a situation where
both the lock exchange process keeps on going (bringing in salt water)
and fresh water is flushed through.

Note that the exact details of this equilibrium are not known. The
current implementation is therefore an approximation based on what
forces may be acting on the fresh and salt water, while making sure that
the approximation holds in the limit of zero flushing discharge and a
flushing discharge equal to or higher than that of the density current.