Fix formulation of equilibrium head when flushing on sea side
We are using the density current formulation to transform the flushing
discharge into a thickness of the fresh boundary layer when the door is
open on the sea side. The velocity in this layer is _not_ the flushing
velocity we calculated before, but a different velocity based on the
flushing discharge and the thickness of the layer. Luckily, the
thickness can still be calculated explicitly from this flushing
discharge and the overall salinity difference, although it does involve
a cubed root call.