Support calling individual phases separately
This commit involves a couple of major changes to accommodate the
calling of the individual phases. Much of the functionality that would
otherwise be duplicated has been split-off into separate functions: -
Functions to perform step 1-4 manually - A function to calculate the
derived parameters

To allow the calling of the individual steps, there was a need for a
"state" of the lock. That way, we can also accommodate someone calling
e.g. phase 3 (levelling) twice in succession, because the lock was
waiting with the doors and valves closed.

To initialize the state, an additional function has been added that
presumes the lock is empty (no ships), at a certain head and with a
certain salinity. All other state variables are calculated from this
assumption and these two input variables.

Note that some functions have been force inlined, to lessen the impact
on performance from splitting the logic into functions. With this forced
inlining there currently is virtually no impact. However, if we did not
force the inlining, the zsf_calculate (looping until steady) function
would be about 20% slower.