Output discharge and salinity for stepwise operation
The discharges and salinities going to/from the lake/sea are typically
needed when running in step-wise operation, e.g. when coupling to
Delft3D (directly, or loosely via time series of discharge + salinity).

Note that steady state calculation already outputs these results, but
they were (incorrectly) assumed not to be necessary for step-wise
operation.