Add input checks and error handling for callers
The externally exposed functions (aside from the default values
function) now return an integer indicating success. If this value is 0,
the function was executed succesfully. If we return a non-zero value,
the corresponding error message can be requested by calling
zsf_error_msg().

Note: This type of error handling will likely be replaced with a
get-last-error-message type. That will allow us to give more information
to the user about what values were wrong, instead of returning a generic
message as we do in this commit.