Rename user-facing sal_* input variables to salinity_*
For many user-facing input variables we already write out the full
quantity name, e.g. head_* and temperature_*. For the salinity, we also
use the full name for the output structs. Using an abbreviation for just
the salinity input variables was therefore an odd one out, which is
fixed in this commit.