Rename 'in/out' volumes, discharges and salinities to 'to/from'
Many times it turned out to be confusing as to what e.g. the flux
"volume_sea_in" would refer to, as it is common to think from the
perspective from the lock. It could then mean water going from lock to
sea (when thinking from the perspective of the sea), or water going from
the sea to the lock (when thinking from the perspective of the lock).
Note that it was supposed to mean the former.

We rename all the in/out variables to use to/from. This way it is
clearer what the direction is, e.g. "volume_to_sea", by always reasoning
from the perspective of the lock.