Fix possible divisions by zero in salinity calculation
There were two divisions by a volume without a guard for zero-volume.
One case where this would result in a division by zero is when flushing
a lot (and no shipping volume), such that there is no volume going from
the lock to the lake (but only the other way).