CI/CD: Add Windows builds
Note that we are using an unstable version of `wheel` to build the
binary package, as we want support for the abi3 tag on Windows. This is
related to https://github.com/pypa/wheel/issues/345 . Note that
installing an abi3 tag on Windows requires pip >= 20.0 (also because
they switched to the `packaging` package to get supported tags).