Excel: Support 64-bit Office
The VBA code has been adjusted such that it either load the functions
from a 64-bit version of "zsf.dll" when dealing with a 64-bit version of
Office. If running a 32-bit version of Office, the 32-bit
"zsf-stdcall.dll" is used to load the functions from.

Note that stdcall is not a thing on x64 platforms, and we can therefore
just use the regular dll.

Also note that for packinging the Excel version of the ZSF, these two
dlls will live next to each other, and next to the Excel xlsm file.