Add option for sill as counter-measure
A sill is 80% effective in reducing the driving height of the density
current. Furthermore, when the door on the fresh side is open, it is
also effective in reducing the amount of volume that can be exchanged.
Currently this is also reduced with 80% of the sill height.

Note that the current implementation is based on first principles and
some early experiments, but no throrough research. The implementation is
therefore is likely to change in the future, but the API should stay
rather similar.