Add option to take distance between door and bubble screen into account
If the bubble screen is placed away from the doors, e.g. outside the
doors to make sure the usable lock length is unaffected by the bubble
screens, the density current will not immediately be reduced in velocity
by the bubble screens. There is a certain volume between the door and
the bubble screen that will be exchanged more quickly.

A positive value means the bubble screen is located _outside_ the lock,
and a negative value will mean it is located _inside_ the lock. A value
of zero (the default) means that the bubble screen is located at the
door itself.

The current implementation assumes that the lock exchange will happen
with the unreduced velocity until it reaches the bubble screen. Note
that this is based on first-principles, and no thorough research has yet
been done. The implementation is therefore likely to change, although
the API will be kept the same as much as possible.