Add ship volume to lake/sea to transport output
When a ship goes through a lock, water is displaced in the opposite
direction. Depending on how a far-field model connected to the ZSF can
model ships, it makes sense to either replace the displaced ship volume
with water, some type of solid/local air pressure field, or flat out
ignore it.

Note that ignoring the ship displacement will lead to translation waves
in many far-field models. Substituting the displaced ship volume with
water is generally a lesser evil.