Docs: Add initial version of docs on the API, theory and examples
The theory part of the documentation is based on a document provided to
the Directorate-General for Public Works and Water Management on the
workings of the ZSF. Said document has been translated to English, with
improvements/corrections made along the way while writing it in
reStructuredText.

Note that the original setup of the ZSF was that of only steady state.
This is somewhat reflected in how the theory documentation is set up in
this commit (and its source document). With the advent of phase-wise
solving, it may make more sense to discuss that first, and consider
steady-state solving a special case to be discussed after. That possible
improvement will be left for a future (set of) commits.