CI/CD: Pin version of `black` used for Python code formatting
A newer version of black, version 20.8b1, would reformat the Python test
code significantly. As newer betas of black might again change the
format, we pin the version until there is a stable release of black.

Note that we do not pin the version of flake8, although that also
sometimes leads to the CI/CD pipeline failing unexpectedly. Generally
though, those updates to flake8 help us catch style/code errors, so we
do prefer to use the last stable version of that for the time being.