Fix salinity_to_sea calculation in Phase 4
With the mass transport to the sea defined as positive, it typically
takes on negative values (as salt moves from the sea to the lake). The
calculation of the salinity seemed to have assumed the mass transport to
be positive.

The mass transport to the sea can be written as as `mt_sea =
salinity_to_sea * volume_to_sea - salinity_from_sea * volume_from_sea`.
Rewriting this to solve for `salinity_to_sea` easily shows that the sign
was wrong.

Note that this output parameter is more like a post-processing parameter
calculated from the mass and volume transports, and that the wrong
calculation therefore did not affect the state (salinity/volume) of the
lock.

Also note that the steady state calculation `zsf_calc_steady` function
calculates the salinity_to_sea in a similar fashion, and had (and still
has) the correct sign.

Lastly, note that most uses of the ZSF involve the salt intrusion on the
fresh/lake side, where the calculation of the respective ingoing and
outgoing salinities were (and still are) correct.